Update README.md
This code is of a car animation. By using c++ language, with a delay time of 10 seconds. The basics of animation is used i.e line, arc, circles etc. In order to make this animation, one must have knowledge of c++ (graphics), understanding of BGI Directory and you are good to go.